feat:添加首页flag列表以及调整flagitem的配色
